requirements.txt
# Core Django Framework
Django==5.2.6
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# API & Authentication  
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
django-cors-headers==4.9.0

# Database
psycopg2-binary==2.9.10
PyMySQL==1.1.2
mysqlclient==2.2.7

# Environment & Production
python-dotenv==1.0.0
gunicorn==22.0.0
whitenoise==6.5.0

# Testing & Quality
pytest==8.4.2
pytest-django==4.11.1
coverage==7.10.7
black==25.9.0
flake8==7.3.0
isort==6.0.1

# Core dependencies
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
uritemplate==4.2.0
inflection==0.5.1
PyJWT==2.10.1
attrs==25.3.0
click==8.3.0
colorama==0.4.6
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytokens==0.1.10
referencing==0.36.2
rpds-py==0.27.1
